Split up ViewController
